requirements.txt
asgiref==3.3.4
bleach==3.3.0
build==0.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docutils==0.17.1
idna==2.10
importlib-metadata==4.5.0
jeepney==0.6.0
keyring==23.0.1
packaging==20.9
pep517==0.10.0
pkginfo==1.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tqdm==4.61.1
twine==3.4.1
urllib3==1.26.4
webencodings==0.5.1
zipp==3.4.1
